+++ http/src/main/scala/com/twitter/finatra/http/exceptions/HttpNackException.scala
@@ -1,0 +1,10 @@
+package com.twitter.finatra.http.exceptions
+
+import scala.util.control.NoStackTrace
+
+class HttpNackException(val retryable: Boolean)
+    extends Exception
+    with NoStackTrace {
+
+  override def getMessage: String = s"HttpNackException(retryable = $retryable)"
+}